% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gprd.R
\name{gprd}
\alias{gprd}
\title{Gaussian Process Regression for Regression Discontinuity}
\usage{
gprd(
  x,
  y,
  cutoff = 0,
  estimator = "piecewise",
  hypers = list(list(b = c(0, 0), B = diag(10, nrow = 2), sigma_y = 1, sigma_f = 1, ell
    = 1), list(b = c(0, 0), B = diag(10, nrow = 2), sigma_y = 1, sigma_f = 1, ell = 1)),
  ci_width = 0.95
)
}
\arguments{
\item{x}{A numeric vector giving the observations of the explanatory variable}

\item{y}{A numeric vector giving the observations of the outcome}

\item{cutoff}{A numeric vector of length one giving the discontinutiy cutoff;
the default is 0.}

\item{estimator}{A character vector of length one giving the estimation
strategy; should be one of "global" (one unknown function for all
observations, with dummy variable added to indicate treatment) or
"piecewise" (different unknown functions for observations to the left of
the cutoff and to the right). The default is "piecewise".}

\item{hypers}{A list giving the hyperparameters of the model. If
\code{estimator = "piecewise"}, should be a list of length two, where each
element is itself a list giving the hyperparameters for the functions to
each side of the cutoff. The list(s) of hypers should have elements
"b" (a numeric vector giving the coefficients' prior mean),
"B" (a numeric matrix giving the coefficients' prior covariance),
"sigma_y" (a numeric vector of length one giving the output variance),
"sigma_f" (a numeric vector of length one giving the function variance),
and "ell" (a numeric vector giving the length scale(s)).}

\item{ci_width}{A numeric vector of length one between 0 and 1 giving
the width of the confidence interval for the treatment effect.}
}
\value{
A list of length five with class \code{gprd} and elements
    \describe{
        \item{tau_mean}{The treatment effect estimate}
        \item{tau_ci}{A named numeric ector of length two giving the
                      requested confidence interval (the names are the
                      confidence interval quantiles)}
        \item{cutoff}{A numeric vector of length one giving the cutoff}
        \item{estimator}{A character vector of length one giving the
                         estimation strategy}
        \item{f}{A list where each element is an object of class \code{gpr}
                 with information on the posterior over the relevant
                 function; if \code{estimator = "global"}, the list will be
                 of length one, giving the posterior over the global mapping
                 from \code{x} to \code{y}, while if
                 \code{estimator = "piecewise"}, the list will be of length two
                 with elements \code{f_l} (giving the posterior over the
                 function to the left of the cutoff) and \code{f_r} (giving
                 the posterior over the function to the right of the cutoff)}
    }

    An object of class \code{gpr} is a list of length eight with the following
elements:

\describe{
    \item{predictive_mean}{The mean of the predictive distribution at test
                           cases \code{test_input}; in this case, the cutoff}
    \item{predictive_var}{The variance of the predictive distribution at test
                          cases \code{test_input}; in this case, the cutoff}
    \item{beta_bar}{The mean of the posterior of the linear mean coefficients}
    \item{training_input}{The input variable obervations for training cases}
    \item{training_outcomes}{The outcome observations for training cases}
    \item{test_input}{The input variable observations predictions were
                      provided for}
    \item{prior_covariance}{The covariance function evaluated for the
                            training inputs, with observation noise added}
    \item{prior_cov_inverse}{The inverse of \code{prior_covariance};
                             having this available is useful for generating
                             other predictions}
    \item{hyperparameters}{The hyperparameters used for the GP regression,
                           provided in a list of length five:
                           \code{beta_prior_cov}, giving the
                           beta prior covariance;
                           \code{beta_prior_mean}, giving the beta prior
                           mean;
                           \code{sigma_f}, giving the scale factor for the
                           GP prior;
                           \code{ell}, giving the length scale for the GP
                           prior;
                           and \code{sigma_y}, giving the outcome observation
                           noise (the likelihood hyper).}
}
}
\description{
Gaussian Process Regression for Regression Discontinuity
}
\seealso{
plot.gprd
}
