% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gprd.R
\name{plot.gprd}
\alias{plot.gprd}
\title{Plot Gaussian Process Regression for Regression Discontinuity}
\usage{
\method{plot}{gprd}(
  x,
  from = "data_min",
  to = "data_max",
  n_points = NULL,
  ci_width = 0.95,
  data_color = "#1c1c1c1c",
  line_color = "black",
  ci_color = "#87878787",
  plot_cutoff = TRUE,
  main_title = "",
  xlab = "",
  ylab = "",
  ...
)
}
\arguments{
\item{x}{A object of class \code{\link{gprd}}}

\item{from}{The lowest x value to plot prediction for.
The default, "data_min", uses the lowest forcing variable observed value.
If both \code{from} and \code{to} are \code{NULL}, only values in the
training data and at the cutoff are plotted.
Otherwise a numeric vector of length one should be given.}

\item{to}{The highest x value to plot prediction for.
The default "data_max" uses the largest forcing variable observed value.
If both \code{from} and \code{to} are \code{NULL}, only values in the
training data and at the cutoff are plotted.
Otherwise a numeric vector of length one should be given.}

\item{n_points}{An integer vector of length one giving the number of
prediction points to plot; if \code{NULL} and \code{from} and \code{to}
are given, \code{n_points = length(seq(from = from, to = to, by = 0.01))}}

\item{ci_width}{A numeric vector of length one between 0 and 1 giving
the width of the confidence interval for tau}

\item{data_color}{A color to plot the data points in}

\item{line_color}{A color to plot the predictive mean line in}

\item{ci_color}{A color to plot the CI polygon in}

\item{plot_cutoff}{A logical vector of length one; if \code{TRUE}, a dashed
vertical line (\code{lty = 2}) marks the cutoff; default is \code{TRUE}}

\item{main_title}{A character vector of length one giving a main title for
the plot}

\item{xlab}{A character vector of length one giving a title for the x axis}

\item{ylab}{A character vector of length one giving a title for the y axis}

\item{...}{Other arguments passed to \code{\link[graphics]{plot}}}
}
\value{
Returns \code{NULL} invisibly
}
\description{
Plot Gaussian Process Regression for Regression Discontinuity
}
\seealso{
gprd
}
