% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotrd.R
\name{plotrd}
\alias{plotrd}
\title{Plot Gaussian Process Regression Discontinuity}
\usage{
plotrd(x, y, cutoff = 0, from = min(x), to = max(x), b_l = c(0, 0),
  b_r = c(0, 0), B_l = diag(10, nrow = 2), B_r = diag(10, nrow = 2),
  output_sigma = 1, cov_sigma = 1, cov_ell = 1, ci_width = 0.95,
  data_color = "#5f5f5f5f", line_color = "black",
  ci_color = "#87878787", main_title = "", xlab = "", ylab = "",
  ...)
}
\arguments{
\item{x}{A numeric vector giving the observations of the explanatory variable}

\item{y}{A numeric vector giving the observations of the outcome}

\item{cutoff}{A numeric vector of length one giving the discontinutiy cutoff;
the default is 0}

\item{from}{A numeric vector of length one giving the lowest x value
to plot prediction for}

\item{to}{A numeric vector of length one giving the highest x value
to plot prediction for}

\item{b_l}{A numeric vector of length two giving the prior means for the
intercept and slope to the left of the cutoff}

\item{b_r}{A numeric vector of length two giving the prior means for the
intercept and slope to the right of the cutoff}

\item{B_l}{A numeric matrix with 2 rows and 2 columns giving the prior
covariance for the intercept and slope to the left of the cutoff}

\item{B_r}{A numeric matrix with 2 rows and 2 columns giving the prior
covariance for the intercept and slope to the right of the cutoff}

\item{output_sigma}{A numeric vector of length one; a hyperparameter for
the observation noise}

\item{cov_sigma}{A numeric vector of length one; a hyperparameter for the
signal noise in the unobserved function}

\item{cov_ell}{A numeric vector of length one; a hyperparameter for the
characteristic length scale for the unobserved function}

\item{ci_width}{A numeric vector of length one between 0 and 1 giving
the width of the confidence interval for tau}

\item{data_color}{A color to plot the data points in}

\item{line_color}{A color to plot the predictive mean line in}

\item{ci_color}{A color to plot the CI polygon in}

\item{main_title}{A character vector of length one giving a main title for
the plot}

\item{xlab}{A character vector of length one giving a title for the x axis}

\item{ylab}{A character vector of length one giving a title for the y axis}

\item{...}{Other arguments passed to \code{\link[graphics]{plot}}}
}
\value{
Returns \code{NULL} invisibly
}
\description{
Plot Gaussian Process Regression Discontinuity
}
