% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.gprd.R
\name{summary.gprd}
\alias{summary.gprd}
\title{Summary of gprd objects}
\usage{
\method{summary}{gprd}(x, digits = 3)
}
\arguments{
\item{x}{A object of class \code{\link{gprd}}}

\item{digits}{An integer vector of length one giving the number of digits
to print; default is 3}
}
\value{
Returns \code{NULL} invisibly
}
\description{
Print the treatment effect estimate and C.I.
}
\seealso{
gprd
}
