% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gprd.R
\name{gprd}
\alias{gprd}
\title{Gaussian Process Regression Discontinuity}
\usage{
gprd(x, y, cutoff = 0, b_l = c(0, 0), b_r = c(0, 0), B_l = diag(10,
  nrow = 2), B_r = diag(10, nrow = 2), output_sigma = 1,
  cov_sigma = 1, cov_ell = 1, ci_width = 0.95)
}
\arguments{
\item{x}{A numeric vector giving the observations of the explanatory variable}

\item{y}{A numeric vector giving the observations of the outcome}

\item{cutoff}{A numeric vector of length one giving the discontinutiy cutoff;
the default is 0.}

\item{b_l}{A numeric vector of length two giving the prior means for the
intercept and slope to the left of the cutoff}

\item{b_r}{A numeric vector of length two giving the prior means for the
intercept and slope to the right of the cutoff}

\item{B_l}{A numeric matrix with 2 rows and 2 columns giving the prior
covariance for the intercept and slope to the left of the cutoff}

\item{B_r}{A numeric matrix with 2 rows and 2 columns giving the prior
covariance for the intercept and slope to the right of the cutoff}

\item{output_sigma}{A numeric vector of length one; a hyperparameter for
the observation noise}

\item{cov_sigma}{A numeric vector of length one; a hyperparameter for the
signal noise in the unobserved function}

\item{cov_ell}{A numeric vector of length one; a hyperparameter for the
characteristic length scale for the unobserved function}

\item{ci_width}{A numeric vector of length one between 0 and 1 giving
the width of the confidence interval for tau}
}
\value{
A named vector of length three with elements
    \describe{
        \item{tau}{The treatment effect estimate}
        \item{tau_low}{The lower bound of the confidence interval}
        \item{tau_high}{The upper bound of the confidence interval}
    }
}
\description{
Gaussian Process Regression Discontinuity
}
