% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gprd.R
\name{gprd}
\alias{gprd}
\title{Gaussian Process Regression Discontinuity}
\usage{
gprd(x, y, cutoff = 0, b_l = c(0, 0), b_r = c(0, 0), B_l = diag(10,
  nrow = 2), B_r = diag(10, nrow = 2), output_sigma = 1,
  cov_sigma = 1, cov_ell = 1, ci_width = 0.95)
}
\arguments{
\item{x}{A numeric vector giving the observations of the explanatory variable}

\item{y}{A numeric vector giving the observations of the outcome}

\item{cutoff}{A numeric vector of length one giving the discontinutiy cutoff;
the default is 0.}

\item{b_l}{A numeric vector of length two giving the prior means for the
intercept and slope to the left of the cutoff}

\item{b_r}{A numeric vector of length two giving the prior means for the
intercept and slope to the right of the cutoff}

\item{B_l}{A numeric matrix with 2 rows and 2 columns giving the prior
covariance for the intercept and slope to the left of the cutoff}

\item{B_r}{A numeric matrix with 2 rows and 2 columns giving the prior
covariance for the intercept and slope to the right of the cutoff}

\item{output_sigma}{A numeric vector of length one; a hyperparameter for
the observation noise}

\item{cov_sigma}{A numeric vector of length one; a hyperparameter for the
signal noise in the unobserved function}

\item{cov_ell}{A numeric vector of length one; a hyperparameter for the
characteristic length scale for the unobserved function}

\item{ci_width}{A numeric vector of length one between 0 and 1 giving
the width of the confidence interval for tau}
}
\value{
A list of length five with class \code{gprd} and elements
    \describe{
        \item{tau_mean}{The treatment effect estimate}
        \item{tau_ci}{A named numeric ector of length two giving the
                      requested confidence interval (the names are the
                      confidence interval quantiles)}
        \item{cutoff}{A numeric vector of length one giving the cutoff}
        \item{f_l}{An object of class \code{gpr} with information on
                   the posterior over the function to the left of the cutoff}
        \item{f_r}{An object of class \code{gpr} with information on
                   the posterior over the function to the right of the cutoff}
    }

    An object of class \code{gpr} is a list of length eight with the following
elements:

\describe{
    \item{predictive_mean}{The mean of the predictive distribution at test
                           cases \code{test_input}; in this case, the cutoff}
    \item{predictive_var}{The variance of the predictive distribution at test
                          cases \code{test_input}; in this case, the cutoff}
    \item{beta_bar}{The mean of the posterior of the linear mean coefficients}
    \item{training_input}{The input variable obervations for training cases}
    \item{training_outcomes}{The outcome observations for training cases}
    \item{test_input}{The input variable observations predictions were
                      provided for}
    \item{prior_covariance}{The covariance function evaluated for the
                            training inputs, with observation noise added}
    \item{prior_cov_inverse}{The inverse of \code{prior_covariance};
                             having this available is useful for generating
                             other predictions}
    \item{hyperparameters}{The hyperparameters used for the GP regression,
                           provided in a list of length five, giving the
                           beta prior covariance, beta prior mean,
                           scale factor for the GP, length scale for the GP,
                           and outcome observation noise.}
}
}
\description{
Gaussian Process Regression Discontinuity
}
\seealso{
plot.gprd
}
