% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_hyperparameters.R
\name{optimize_hyperparameters}
\alias{optimize_hyperparameters}
\title{Optimize hyperparameters for GPRD}
\usage{
optimize_hyperparameters(
  x,
  y,
  cutoff = 0,
  estimator = "piecewise",
  b = NULL,
  B = NULL,
  method = "CG",
  control = list(type = 2),
  ...
)
}
\arguments{
\item{x}{A numeric vector; the forcing variable}

\item{y}{A numeric vector; the outcome variable}

\item{cutoff}{A numeric vector of length one; the treatment cutoff
(default is 0)}

\item{estimator}{A character vector of length one; should be one of "global"
or "piecewise". The default is "global".}

\item{b}{A numeric vector giving the prior mean for the coefficients for the
mean function. If \code{NULL} (the default), all coefficients have a
prior mean of 0.}

\item{B}{A numeric matrix giving the prior covariance for the coefficients
for the mean function. If \code{NULL} (the default), the prior
covariance for the coefficients is a diagonal matrix, with the diagonal
elements set equal to 10.}

\item{method}{A character vector of length one giving the optimization
routine to use; default is "CG", the conjugate gradient method.
See \code{\link[stats]{optim}} for other options.}

\item{control}{A list of control parameters for optim}

\item{...}{Other arguments passed to \code{\link[stats]{optim}}}
}
\description{
Optimize hyperparameters for GPRD
}
\seealso{
gprd
}
